arranged code with coding standards
