+++ go.mod
@@ -3,7 +3,15 @@ module serial-data-decryptor
 go 1.18
 
 require (
+	github.com/golang/gddo v0.0.0-20210115222349-20d68f94ee1f
 	github.com/joho/godotenv v1.4.0
 	github.com/sirupsen/logrus v1.9.0
+)
+
+require (
+	github.com/go-stack/stack v1.6.0 // indirect
+	github.com/inconshreveable/log15 v0.0.0-20170622235902-74a0988b5f80 // indirect
+	github.com/mattn/go-colorable v0.0.10-0.20170816031813-ad5389df28cd // indirect
+	github.com/mattn/go-isatty v0.0.2 // indirect
 	golang.org/x/sys v0.0.0-20220715151400-c0bba94af5f8 // indirect
 )
