+++ models/data.go
@@ -1,0 +1,3 @@
+package models
+
+type Data map[string]interface{}
