+++ utility/env.go
@@ -1,0 +1,14 @@
+package utility
+
+import (
+	"log"
+	"os"
+)
+
+func GetEnvAsserted(envVarName string) string {
+	var thisEnvVar = os.Getenv(envVarName)
+	if len(thisEnvVar) == 0 {
+		log.Fatal(envVarName, " was not found in the current environment")
+	}
+	return thisEnvVar
+}
